file updates for better readability. separated ECC into file types and components
